BrainWaveMonitor1.0.cs
BrainWaveMonitor1.0.cs: This code is written in the C# programming language to create a brainwave monitor application using the ScottPlot library. The application visualizes incoming data live within a graphical window. The application window includes a ScottPlot graphic object and a timer. Upon starting the application, the timer adds new data points to the graph object at specified intervals and removes old points to keep the graph continuously updated with live data. The application provides a user interface with buttons for starting and stopping the data flow.

BrainWaveMonitor1.0.1.cs; This program utilizes the ScottPlot library to generate an app for monitoring brain waves. The app displays real-time data by establishing a connection to a device. The app interface comprises the ScottPlot graph along with a few other controls such as a "Connect" button and a "Disconnect" button. Upon clicking the "Connect" button, the application links up with the designated device and starts exhibiting live data. The "Disconnect" button terminates the device connection and halts the data flow.